restapi/requirements.txt
Django==2.2
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt
Pillow==5.2.0
PyJWT==1.6.4
pytz==2018.5
six==1.11.0
psycopg2==2.8.6
celery
redis
